requirements.txt
langchain-couchbase==0.2.0
langchain-openai==0.2.5
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.39.0
tqdm==4.66.6
openai==1.54.4
